drop py2 and compat module
