+++ pyproject.toml
@@ -5,6 +5,7 @@ build-backend = "setuptools.build_meta"
 [tool.black]
 line-length = 79
 skip-string-normalization = true
+target_version = ["py37"]
 exclude = '''
 /(
     \.git
