+++ rtd.txt
@@ -1,1 +1,1 @@
-.[docs]
+-e .[docs]
