+++ setup.cfg
@@ -2,7 +2,7 @@
 universal = 1
 
 [metadata]
-license_file = LICENSE.txt
+license_files = LICENSE.txt
 
 [check-manifest]
 ignore =
