+++ setup.py
@@ -20,6 +20,7 @@ setup(
         'Integrated process monitor for developing and reloading daemons.'
     ),
     long_description=readme + '\n\n' + changes,
+    long_description_content_type='text/x-rst',
     author='Michael Merickel',
     author_email='pylons-discuss@googlegroups.com',
     url='https://github.com/Pylons/hupper',
@@ -27,7 +28,7 @@ setup(
     packages=find_packages('src', exclude=['tests']),
     package_dir={'': 'src'},
     include_package_data=True,
-    python_requires='>=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*',
+    python_requires='>=3.7',
     extras_require={'docs': docs_require, 'testing': tests_require},
     entry_points={"console_scripts": ["hupper = hupper.cli:main"]},
     zip_safe=False,
@@ -37,8 +38,6 @@ setup(
         'Intended Audience :: Developers',
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
-        'Programming Language :: Python :: 2',
-        'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
