+++ src/hupper/interfaces.py
@@ -1,19 +1,18 @@
-import abc
-
-from .compat import with_metaclass
+from abc import ABC, abstractmethod
 
 
-class IReloaderProxy(with_metaclass(abc.ABCMeta)):
-    @abc.abstractmethod
+class IReloaderProxy(ABC):
+    @abstractmethod
     def watch_files(self, files):
         """Signal to the monitor to track some custom paths."""
 
-    @abc.abstractmethod
+    @abstractmethod
     def trigger_reload(self):
         """Signal the monitor to execute a reload."""
 
 
-class IFileMonitorFactory(with_metaclass(abc.ABCMeta)):
+class IFileMonitorFactory(ABC):
+    @abstractmethod
     def __call__(self, callback, **kw):
         """Return an :class:`.IFileMonitor` instance.
 
@@ -32,16 +31,16 @@ class IFileMonitorFactory(with_metaclass(abc.ABCMeta)):
         """
 
 
-class IFileMonitor(with_metaclass(abc.ABCMeta)):
-    @abc.abstractmethod
+class IFileMonitor(ABC):
+    @abstractmethod
     def add_path(self, path):
         """Start monitoring a new path."""
 
-    @abc.abstractmethod
+    @abstractmethod
     def start(self):
         """Start the monitor. This method should not block."""
 
-    @abc.abstractmethod
+    @abstractmethod
     def stop(self):
         """Trigger the monitor to stop.
 
@@ -49,20 +48,20 @@ class IFileMonitor(with_metaclass(abc.ABCMeta)):
 
         """
 
-    @abc.abstractmethod
+    @abstractmethod
     def join(self):
         """Block until the monitor has stopped."""
 
 
-class ILogger(with_metaclass(abc.ABCMeta)):
-    @abc.abstractmethod
+class ILogger(ABC):
+    @abstractmethod
     def error(self, msg):
         """Record an error message."""
 
-    @abc.abstractmethod
+    @abstractmethod
     def info(self, msg):
         """Record an informational message."""
 
-    @abc.abstractmethod
+    @abstractmethod
     def debug(self, msg):
         """Record a debug-only message."""
