+++ src/hupper/polling.py
@@ -58,6 +58,6 @@ def get_mtime(path):
         stat = os.stat(path)
         if stat:
             return stat.st_mtime
-    except (OSError, IOError):  # pragma: no cover
+    except OSError:  # pragma: no cover
         pass
     return 0
