+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist =
     lint,
-    py27,py37,py38,py39,py310,py311,pypy,pypy3,
+    py37,py38,py39,py310,py311,pypy3,
     docs,coverage
 
 isolated_build = true
@@ -38,19 +38,24 @@ extras =
     docs
 
 [testenv:lint]
-skip_install = true
+skip_install = True
 commands =
-    flake8 src/hupper/
-    isort --check src/hupper tests setup.py
+    isort --check-only --df src/hupper tests setup.py
     black --check --diff src/hupper tests setup.py
-    python setup.py check -r -s -m
+    flake8 src/hupper tests setup.py
     check-manifest
+    # build sdist/wheel
+    python -m build .
+    twine check dist/*
 deps =
     black
+    build
     check-manifest
     flake8
+    flake8-bugbear
     isort
     readme_renderer
+    twine
 
 [testenv:format]
 skip_install = true
@@ -65,12 +70,16 @@ deps =
 skip_install = true
 commands =
     # clean up build/ and dist/ folders
-    python -c 'import shutil; shutil.rmtree("dist", ignore_errors=True)'
-    python setup.py clean --all
-    # build sdist
-    python setup.py sdist --dist-dir {toxinidir}/dist
-    # build wheel from sdist
-    pip wheel -v --no-deps --no-index --no-build-isolation --wheel-dir {toxinidir}/dist --find-links {toxinidir}/dist hupper
+    python -c 'import shutil; shutil.rmtree("build", ignore_errors=True)'
+    # Make sure we aren't forgetting anything
+    check-manifest
+    # build sdist/wheel
+    python -m build .
+    # Verify all is well
+    twine check dist/*
+
 deps =
-    setuptools
-    wheel
+    build
+    check-manifest
+    readme_renderer
+    twine
